[ REFACTOR ] 권한 검증 메서드 서비스 계층으로 이동 ( #32 )
