+++ com.first-class.msa.hub/src/main/java/com/first_class/msa/hub/presentation/controller/HubController.java
@@ -3,7 +3,6 @@ package com.first_class.msa.hub.presentation.controller;
 import com.first_class.msa.hub.application.dto.ResDTO;
 import com.first_class.msa.hub.application.dto.ResHubPostDTO;
 import com.first_class.msa.hub.application.dto.ResHubSearchDTO;
-import com.first_class.msa.hub.application.service.AuthService;
 import com.first_class.msa.hub.application.service.HubService;
 import com.first_class.msa.hub.domain.model.Hub;
 import com.first_class.msa.hub.presentation.request.ReqHubPostDTO;
@@ -19,28 +18,23 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.Objects;
-
 @RestController
 @RequiredArgsConstructor
 public class HubController {
 
     private final HubService hubService;
-    private final AuthService authService;
 
     @PostMapping("/hubs")
     public ResponseEntity<ResDTO<ResHubPostDTO>> postBy(@RequestHeader("X-User-Id") Long userId,
                                                         @RequestHeader("X-User-Account") String account,
                                                         @Valid @RequestBody ReqHubPostDTO req) {
 
-        validateUserRole(userId);
-
 
         return new ResponseEntity<>(
                 ResDTO.<ResHubPostDTO>builder()
                         .code(HttpStatus.OK.value())
                         .message("허브 생성에 성공하였습니다.")
-                        .data(hubService.postBy(account, req))
+                        .data(hubService.postBy(userId, account, req))
                         .build(),
                 HttpStatus.OK
         );
@@ -67,11 +61,10 @@ public class HubController {
                                                 @PathVariable(name = "hubId") Long hubId,
                                                 @Valid @RequestBody ReqHubPutByIdDTO dto) {
 
-        validateUserRole(userId);
         // --
         // TODO : DataIntegrityViolationException 예외처리
         // --
-        hubService.putBy(account, hubId, dto);
+        hubService.putBy(userId, account, hubId, dto);
 
         return new ResponseEntity<>(
                 ResDTO.builder()
@@ -87,9 +80,7 @@ public class HubController {
                                                    @RequestHeader("X-User-Account") String account,
                                                    @PathVariable(name = "hubId") Long hubId) {
 
-        validateUserRole(userId);
-
-        hubService.deleteBy(account, hubId);
+        hubService.deleteBy(userId, account, hubId);
 
         return new ResponseEntity<>(
                 ResDTO.builder()
@@ -98,11 +89,5 @@ public class HubController {
                         .build(),
                 HttpStatus.OK
         );
-    }
-
-    private void validateUserRole(Long userId) {
-        if(!Objects.equals("MASTER", authService.getRoleBy(userId).getRole())) {
-            throw new IllegalArgumentException("접근 권한이 없습니다.");
-        }
     }
 }
